requirements.txt
alembic==1.13.0
bcrypt==4.1.1
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
flask-mailman==1.0.0
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
httplib2==0.22.0
idna==3.6
importlib-resources==6.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mkdocs-material-extensions==1.3.1
mypy==1.7.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
passlib==1.7.4
pip==23.3.1
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.0.2
six==1.16.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
types-Flask-Migrate==4.0.0.7
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
